Switch to hardware oriented data types

+++ inc/EFM8BB52/delay.h
@@ -1,10 +1,12 @@
 #ifndef _EFM8BB52_DELAY_H_
 #define _EFM8BB52_DELAY_H_
 
+#include <EFM8BB52/types.h>
+
 /* all methods here will reconfigure and use Timer 5 */
 
-void delay(unsigned int seconds);
-void delay_ms(unsigned int milliseconds);
-void delay_us(unsigned long microseconds);
+void delay(u16 seconds);
+void delay_ms(u16 milliseconds);
+void delay_us(u32 microseconds);
 
 #endif
